Changed LogicSimulation

+++ src/main/java/org/example/models/Herbivore.java
@@ -18,20 +18,46 @@ public class Herbivore extends Creature {
 
     @Override
     public void makeMove(GameMap gameMap) {
-        Entity target=coordinateService.findTarget(this, gameMap.getGameMap());
+
+        if (!this.isAlive()){
+//            System.out.println("Herbivore is dead");
+            return;
+        }
+
+        Entity target = coordinateService.findTarget(this, gameMap.getGameMap());
         if (target != null) {
-            List<Coordinates> path=coordinateService.getShortPath(this, target);
-            if (!path.isEmpty()) {
-                for (Coordinates nextStep : path) {
-                    if (gameMap.isSquareEmpty(nextStep)) {
-                        gameMap.deleteEntity(this);
-                        this.setCoordinates(nextStep);
-                        gameMap.setStaticObjects(nextStep, this);
-                        break;
+//            System.out.println("Травоядное видит цель: " + target.getClass().getSimpleName() + " на " + target.getCoordinates());
+            if (isAdjacent(target.getCoordinates())){
+//                System.out.println("Травоядное на клетке с травой!");
+                if (target instanceof Grass) {
+                    gameMap.deleteEntity(target);
+                    this.health += 2;
+//                    System.out.println("Травоядное съело траву. Здоровье: " + this.health);
+                }
+            } else {
+//                System.out.println("Травоядное двигается к траве...");
+                List<Coordinates> path = coordinateService.getShortPath(this, target);
+                if (!path.isEmpty()) {
+                    for (Coordinates nextStep : path) {
+                        if (gameMap.isSquareEmpty(nextStep)) {
+                            gameMap.deleteEntity(this);
+                            this.setCoordinates(nextStep);
+                            gameMap.setStaticObjects(nextStep, this);
+//                            System.out.println("Травоядное переместилось на " + nextStep);
+                            break;
+                        }
                     }
                 }
             }
+        }else {
+//            System.out.println("Травоядное не видит травы.");
         }
+    }
+
+    private boolean isAdjacent(Coordinates target) {
+        int dx = Math.abs(this.getCoordinates().x - target.x);
+        int dy = Math.abs(this.getCoordinates().y - target.y);
+        return (dx == 0 && dy == 1) || (dx == 1 && dy == 0);
     }
 }
